pixelformats: Add memory_alignment class parameter
Pixels are aligned in memory. Even if this alignment is platform
dependent, many platforms align on 16 bytes. Make this value the default
one but allow PixelFormat to be instantiate on plaforms where alignement
is different.

Tested with Bayer formats on Rpi5 and RB5 (were pixels are aligned on 16
bytes) and Rpi4 (were pixels are aligned on 32 bytes).

While at it, add (missing) typing for the 'align' parameter in the
'framesize' and 'dumb_size' functions.

Signed-off-by: Sylvain Petinot <sylvain.petinot@foss.st.com>